# Benchmark test for GetFullPath with redundant segments