--- a/src/benchmarks/micro/libraries/System.Runtime.Extensions/Perf.Path.cs
+++ b/src/benchmarks/micro/libraries/System.Runtime.Extensions/Perf.Path.cs
@@ -16,6 +16,7 @@ namespace System.IO.Tests
         private readonly string _testPath500 = PerfUtils.CreateString(500);
         private readonly string _testPath1000 = PerfUtils.CreateString(1000);
         private readonly string _testPathNoRedundantSegments = "/home/user/runtime/src/coreclr/runtime/src/libraries/System.Private.CoreLib/src/System/IO/Path.cs";
+        private readonly string _testPathWithRedundantSegments = "/home/user/runtime/src/coreclr/runtime/src/libraries/System.Private.CoreLib/src/System/IO/..//./Path.cs";
 
         [Benchmark]
         public string Combine() => Path.Combine(_testPath, _testPath10);
@@ -48,6 +49,9 @@ namespace System.IO.Tests
 
         [Benchmark]
         public void GetFullPathNoRedundantSegments() => Path.GetFullPath(_testPathNoRedundantSegments);
+
+        [Benchmark]
+        public void GetFullPathWithRedundantSegments() => Path.GetFullPath(_testPathWithRedundantSegments);
 
         [Benchmark]
         public string GetPathRoot() => Path.GetPathRoot(_testPath);