--- a/src/benchmarks/micro/libraries/System.Runtime.Extensions/Perf.Path.cs
+++ b/src/benchmarks/micro/libraries/System.Runtime.Extensions/Perf.Path.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System.Runtime.InteropServices;
 using BenchmarkDotNet.Attributes;
 using MicroBenchmarks;
 
@@ -15,8 +16,26 @@ namespace System.IO.Tests
         private readonly string _testPath200 = PerfUtils.CreateString(200);
         private readonly string _testPath500 = PerfUtils.CreateString(500);
         private readonly string _testPath1000 = PerfUtils.CreateString(1000);
-        private readonly string _testPathNoRedundantSegments = "/home/user/runtime/src/coreclr/runtime/src/libraries/System.Private.CoreLib/src/System/IO/Path.cs";
-        private readonly string _testPathWithRedundantSegments = "/home/user/runtime/src/coreclr/runtime/src/libraries/System.Private.CoreLib/src/System/IO/..//./Path.cs";
+        private string _testPathNoRedundantSegments;
+        private string _testPathWithRedundantSegments;
+
+        [GlobalSetup]
+        public void SetupPaths()
+        {
+            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
+            {
+                // This fully qualified path will be normalized by the Windows P/Invoke
+                _testPathNoRedundantSegments = @"C:\repos\runtime\src\coreclr\runtime\src\libraries\System.Private.CoreLib\src\System\IO\Path.cs";
+                // This unqualified path will be analyzed by our RedundantSegments.Windows code
+                _testPathWithRedundantSegments = @"runtime\src\coreclr\runtime\src\libraries\System.Private.CoreLib\src\System\IO\Extra\..\Path.cs";
+            }
+            else
+            {
+                // Both paths will be analyzed by our RedundantSegments.Unix code
+                _testPathNoRedundantSegments = "/home/user/runtime/src/coreclr/runtime/src/libraries/System.Private.CoreLib/src/System/IO/Path.cs";
+                _testPathWithRedundantSegments = "runtime/src/coreclr/runtime/src/libraries/System.Private.CoreLib/src/System/IO/Extra/../Path.cs";
+            }
+        }
 
         [Benchmark]
         public string Combine() => Path.Combine(_testPath, _testPath10);