@billwert 2020-07-13
I think this test should have more typical cases compared to these edge cases. (Or at least include them.) I suspect `..\.\..\.\.\folder1\.\toremove\.\.\..\folder2\.\toremove\toremove\..\.\.\..\.\folder3\.\.\....\.\.\..\\.` is not a very common path.