@billwert 2020-07-14
One other thought about these tests. It looks like the expected improvement is ~2%, right? In general our system isn't going to flag that as a regression if we lost this optimization somehow. That's typically below the noise threshold we hold. @DrewScoggins can opine some here.