@carlossanlop 2021-01-26
@adamsitnik @Jozkee @jkotas I ran these benchmarks, and they are giving me a different result than what we saw with the `Stopwatch` test in https://github.com/dotnet/runtime/pull/37939#issuecomment-763188858 . Would you mind taking a look?