@carlossanlop 2021-01-27
Updated the description to include Windows benchmark tests too.